+++ ChangeLog.md
@@ -9,15 +9,15 @@ All notable changes to this project will be documented in this file.
 
 ### Added 
 
-- functions to handle locks based on a semphore
+- functions to handle locks based on a semaphore
 
 ### Fixed 
 
 - Venice to follow the Java rules when propagating exceptions from try-with-resources
   blocks:
-    1. exception from body block
+    1. exception from finally block
     2. exception from catch block
-    3. exception from finally block
+    3. exception from body block
     4. exception from resource auto-close
 
 ### Updated